requirements.txt
asgiref==3.6.0
channels==4.0.0
Django==4.1.4
djangochannelsrestframework==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.7
sqlparse==0.4.3
